test: split PDF internal destination cases
